+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Fixed
 * <How do the end-user experience this issue? what was the impact?> ([#????](https://github.com/realm/realm-core/issues/????), since v?.?.?)
 * Deleting an object in an asymmetric table would cause a crash. Likely to solve [#1537](https://github.com/realm/realm-kotlin/issues/1537), since v12.1.0.
+* Fixed FLX subscriptions not being fullfilled if the session was interrupted during bootstrapping. ([PR 7042](https://github.com/realm/realm-core/pull/7042))
 
 ### Breaking changes
 * None.